+++ Sources/BraintreeDropIn/Custom Views/BTDropInPaymentSelectionCell.h
@@ -1,6 +1,6 @@
 #import <UIKit/UIKit.h>
 #import "BTUIKPaymentOptionCardView.h"
-#import <BraintreeDropIn/BTUIKPaymentOptionType.h>
+#import <BraintreeDropIn/BTDropInPaymentMethodType.h>
 
 @interface BTDropInPaymentSelectionCell : UITableViewCell
 
@@ -8,6 +8,6 @@
 @property (nonatomic, strong) UILabel* detailLabel;
 @property (nonatomic, strong) BTUIKPaymentOptionCardView* iconView;
 @property (nonatomic, strong) UIView *bottomBorder;
-@property (nonatomic) BTUIKPaymentOptionType type;
+@property (nonatomic) BTDropInPaymentMethodType type;
 
 @end