+++ CHANGELOG.md
@@ -17,6 +17,9 @@
   * Venmo
     * Add `venmoRequest` to `BTDropInRequest`
     * Remove `vaultVenmo` property from `BTDropInRequest` (use `venmoRequest.vault` instead)
+  * BTDropInResult
+    * Rename `BTUIKPaymentOptionType` enum to `BTDropInPaymentMethodType`
+    * Rename `paymentOptionType` property to `paymentMethodType`
 
 ## 9.0.0-beta1 (2021-03-29)